chore(streams): clean up unused-import warnings in new code
Removes the _keep / _keepGen / _keepClose / _keepInt16 stubs I
sprinkled in to silence -Wunused-imports during the parity
sweep, plus the matching unused imports themselves:

  * Kafka.Streams.Discovery.Subscription:   drop Int16/Int32 import
  * Kafka.Streams.Discovery.RemoteIQ:        drop unStoreName import
  * Kafka.Streams.Query:                     drop Set / unused store-import
  * Kafka.Streams.DSL.Repartitioned:         drop ByteString import
  * Kafka.Streams.State.Window.Timestamped:  drop Generic import
  * Kafka.Streams.Runtime: drop redundant fromIntegral (-Widentities)

The pre-existing -Wunused-imports across the streams tree
(Topology.hs, Mock/*, TopologyDescription.hs etc.) are
unchanged; those are scaffolding I shouldn't touch without
audit.

All 369 streams tests still green.

+++ wireform-kafka/streams/src/Kafka/Streams/Runtime.hs
@@ -550,7 +550,7 @@ eventLoop ks driver engine = go
                     else HashMap.fromListWith max
                            [ ( KC.TopicPartition
                                  (KC.crTopic rec)
-                                 (fromIntegral (KC.crPartition rec))
+                                 (KC.crPartition rec)
                              , KC.crOffset rec + 1
                              )
                            | rec <- recs
@@ -637,7 +637,7 @@ multiEventLoop ks driver pool = go
                     else HashMap.fromListWith max
                            [ ( KC.TopicPartition
                                  (KC.crTopic rec)
-                                 (fromIntegral (KC.crPartition rec))
+                                 (KC.crPartition rec)
                              , KC.crOffset rec + 1
                              )
                            | rec <- recs
